Quicksilver release version 11.1.0

+++ Sources/EPiServer.Reference.Commerce.Site/Infrastructure/ImportSiteContent.cs
@@ -107,7 +107,11 @@ namespace EPiServer.Reference.Commerce.Site.Infrastructure
         {
             var warehouseRepository = ServiceLocator.Current.GetInstance<IWarehouseRepository>();
             var defaultWarehouse = warehouseRepository.Get("default");
-            warehouseRepository.Delete(defaultWarehouse.WarehouseId.Value);
+
+            if (defaultWarehouse != null && defaultWarehouse.WarehouseId.HasValue)
+            {
+                warehouseRepository.Delete(defaultWarehouse.WarehouseId.Value);
+            }
         }
 
         private void EnableCurrencies()
